+++ app/src/main/java/com/android/quizcafe/core/designsystem/TextField.kt
@@ -1,0 +1,85 @@
+package com.android.quizcafe.core.designsystem
+
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.android.quizcafe.core.designsystem.theme.Typography
+import com.android.quizcafe.core.designsystem.theme.errorLight
+import com.android.quizcafe.core.designsystem.theme.primaryContainerLight
+import com.android.quizcafe.core.designsystem.theme.primaryLight
+import com.android.quizcafe.core.designsystem.theme.surfaceBrightLight
+
+@Composable
+fun QuizCafeTextField(
+    value: String,
+    onValueChange: (String) -> Unit,
+    modifier: Modifier = Modifier,
+    enabled: Boolean = true,
+    label: String? = null,
+    isPassword: Boolean = false,
+    errorMessage: String? = null,
+    keyboardOptions: KeyboardOptions = KeyboardOptions.Default
+) {
+    Column(modifier = Modifier.fillMaxWidth()) {
+        if (label != null) {
+            Text(label)
+            Spacer(modifier = Modifier.height(4.dp))
+        }
+
+        TextField(
+            value = value,
+            onValueChange = onValueChange,
+            visualTransformation = if (isPassword) PasswordVisualTransformation() else VisualTransformation.None,
+            modifier = if (errorMessage != null) modifier.border(
+                1.dp,
+                Color.Red,
+                RoundedCornerShape(10.dp)
+            ) else modifier,
+            enabled = enabled,
+            colors = TextFieldDefaults.colors(
+                focusedTextColor = Color.Black,
+                unfocusedTextColor = Color.DarkGray,
+                disabledContainerColor = surfaceBrightLight,
+                focusedContainerColor = primaryContainerLight,
+                unfocusedContainerColor = primaryContainerLight,
+                focusedIndicatorColor = Color.Transparent,
+                unfocusedIndicatorColor = Color.Transparent,
+                disabledIndicatorColor = Color.Transparent,
+                cursorColor = primaryLight
+            ),
+            shape = RoundedCornerShape(8.dp),
+            singleLine = true,
+            keyboardOptions = keyboardOptions,
+            textStyle = TextStyle(fontSize = 16.sp)
+        )
+
+        if (errorMessage != null) {
+            Spacer(Modifier.height(4.dp))
+            Text(
+                text = errorMessage,
+                modifier = Modifier
+                    .align(Alignment.End)
+                    .padding(4.dp),
+                color = errorLight,
+                style = Typography.labelSmall,
+            )
+        }
+    }
+}
