+++ app/src/main/java/com/android/quizcafe/feature/signup/SignUpScreen.kt
@@ -44,9 +44,9 @@ import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.android.quizcafe.R
 import com.android.quizcafe.core.designsystem.QuizCafeButton
+import com.android.quizcafe.core.designsystem.QuizCafeTextField
 import com.android.quizcafe.core.ui.QuizCafeTopAppBar
 import com.android.quizcafe.core.ui.TopAppBarTitle
-import com.android.quizcafe.feature.login.QuizCafeTextField
 import kotlinx.coroutines.delay
 
 @OptIn(ExperimentalMaterial3Api::class)
